test(security): align hardened fixture coverage
