+++ crates/tandem-server/src/http/tests/mod.rs
@@ -77,6 +77,55 @@ use crate::http::global::sanitize_relative_subpath;
 
 pub(super) use crate::test_support::{next_event_of_type, test_state};
 
+struct FixedCompletionProvider {
+    response: String,
+}
+
+#[async_trait::async_trait]
+impl tandem_providers::Provider for FixedCompletionProvider {
+    fn info(&self) -> tandem_types::ProviderInfo {
+        tandem_types::ProviderInfo {
+            id: "fixed-completion-test".to_string(),
+            name: "Fixed Completion Test".to_string(),
+            models: vec![tandem_types::ModelInfo {
+                id: "fixed-completion-test-1".to_string(),
+                provider_id: "fixed-completion-test".to_string(),
+                display_name: "Fixed Completion Test 1".to_string(),
+                context_window: 8_192,
+            }],
+        }
+    }
+
+    async fn complete(
+        &self,
+        _prompt: &str,
+        _model_override: Option<&str>,
+    ) -> anyhow::Result<String> {
+        Ok(self.response.clone())
+    }
+
+    async fn complete_with_auth_override(
+        &self,
+        _prompt: &str,
+        _model_override: Option<&str>,
+        _auth_override: tandem_providers::ProviderAuthOverride,
+    ) -> anyhow::Result<String> {
+        Ok(self.response.clone())
+    }
+}
+
+pub(super) async fn install_fixed_completion_provider(state: &AppState, response: &str) {
+    state
+        .providers
+        .replace_for_test(
+            vec![Arc::new(FixedCompletionProvider {
+                response: response.to_string(),
+            })],
+            Some("fixed-completion-test".to_string()),
+        )
+        .await;
+}
+
 pub(super) fn write_pack_zip(path: &std::path::Path, manifest: &str) {
     write_pack_zip_with_entries(path, manifest, &[("README.md", "# pack")]);
 }
@@ -98,6 +147,81 @@ pub(super) fn write_pack_zip_with_entries(
         std::io::Write::write_all(&mut zip, body.as_bytes()).expect("write extra entry");
     }
     zip.finish().expect("finish zip");
+}
+
+pub(super) struct TrustedPackKeyGuard {
+    previous: Option<String>,
+}
+
+impl Drop for TrustedPackKeyGuard {
+    fn drop(&mut self) {
+        if let Some(previous) = self.previous.as_deref() {
+            std::env::set_var("TANDEM_PACK_TRUSTED_PUBLIC_KEYS", previous);
+        } else {
+            std::env::remove_var("TANDEM_PACK_TRUSTED_PUBLIC_KEYS");
+        }
+    }
+}
+
+pub(super) fn write_signed_pack_zip(path: &std::path::Path, manifest: &str) -> TrustedPackKeyGuard {
+    write_signed_pack_zip_with_entries(path, manifest, &[("README.md", "# pack")])
+}
+
+pub(super) fn write_signed_pack_zip_with_entries(
+    path: &std::path::Path,
+    manifest: &str,
+    extra_entries: &[(&str, &str)],
+) -> TrustedPackKeyGuard {
+    use base64::Engine;
+    use ed25519_dalek::{Signer, SigningKey};
+    use sha2::{Digest, Sha256};
+
+    let mut entries = vec![("tandempack.yaml", manifest)];
+    entries.extend_from_slice(extra_entries);
+    let mut ordered = entries
+        .iter()
+        .map(|(name, body)| ((*name).to_string(), body.as_bytes().to_vec()))
+        .collect::<Vec<_>>();
+    ordered.sort_by(|left, right| left.0.cmp(&right.0));
+    let mut hasher = Sha256::new();
+    for (name, body) in &ordered {
+        hasher.update((name.len() as u64).to_be_bytes());
+        hasher.update(name.as_bytes());
+        hasher.update((body.len() as u64).to_be_bytes());
+        hasher.update(body);
+    }
+    let digest: [u8; 32] = hasher.finalize().into();
+    let signing_key = SigningKey::from_bytes(&[11u8; 32]);
+    let signature = signing_key.sign(&digest);
+    let envelope = json!({
+        "key_id": "http-test-publisher",
+        "signature": base64::engine::general_purpose::STANDARD.encode(signature.to_bytes()),
+    })
+    .to_string();
+
+    let file = std::fs::File::create(path).expect("create signed zip");
+    let mut zip = zip::ZipWriter::new(file);
+    let opts = zip::write::SimpleFileOptions::default()
+        .compression_method(zip::CompressionMethod::Deflated);
+    for (name, body) in entries {
+        zip.start_file(name, opts).expect("start signed entry");
+        std::io::Write::write_all(&mut zip, body.as_bytes()).expect("write signed entry");
+    }
+    zip.start_file("tandempack.sig", opts)
+        .expect("start pack signature");
+    std::io::Write::write_all(&mut zip, envelope.as_bytes()).expect("write pack signature");
+    zip.finish().expect("finish signed zip");
+
+    let previous = std::env::var("TANDEM_PACK_TRUSTED_PUBLIC_KEYS").ok();
+    std::env::set_var(
+        "TANDEM_PACK_TRUSTED_PUBLIC_KEYS",
+        format!(
+            "http-test-publisher={}",
+            base64::engine::general_purpose::STANDARD
+                .encode(signing_key.verifying_key().to_bytes())
+        ),
+    );
+    TrustedPackKeyGuard { previous }
 }
 
 pub(super) fn write_plain_zip_without_marker(path: &std::path::Path) {
