+++ crates/tandem-server/src/http/tests/packs.rs
@@ -2,80 +2,6 @@
 // Licensed under the Business Source License 1.1
 
 use super::*;
-use base64::Engine;
-use ed25519_dalek::{Signer, SigningKey};
-use sha2::{Digest, Sha256};
-
-struct TrustedPackKeyGuard {
-    previous: Option<String>,
-}
-
-impl Drop for TrustedPackKeyGuard {
-    fn drop(&mut self) {
-        if let Some(previous) = self.previous.as_deref() {
-            std::env::set_var("TANDEM_PACK_TRUSTED_PUBLIC_KEYS", previous);
-        } else {
-            std::env::remove_var("TANDEM_PACK_TRUSTED_PUBLIC_KEYS");
-        }
-    }
-}
-
-fn write_signed_pack_zip(path: &std::path::Path, manifest: &str) -> TrustedPackKeyGuard {
-    write_signed_pack_zip_with_entries(path, manifest, &[("README.md", "# pack")])
-}
-
-fn write_signed_pack_zip_with_entries(
-    path: &std::path::Path,
-    manifest: &str,
-    extra_entries: &[(&str, &str)],
-) -> TrustedPackKeyGuard {
-    let mut entries = vec![("tandempack.yaml", manifest)];
-    entries.extend_from_slice(extra_entries);
-    let mut ordered = entries
-        .iter()
-        .map(|(name, body)| ((*name).to_string(), body.as_bytes().to_vec()))
-        .collect::<Vec<_>>();
-    ordered.sort_by(|left, right| left.0.cmp(&right.0));
-    let mut hasher = Sha256::new();
-    for (name, body) in &ordered {
-        hasher.update((name.len() as u64).to_be_bytes());
-        hasher.update(name.as_bytes());
-        hasher.update((body.len() as u64).to_be_bytes());
-        hasher.update(body);
-    }
-    let digest: [u8; 32] = hasher.finalize().into();
-    let signing_key = SigningKey::from_bytes(&[11u8; 32]);
-    let signature = signing_key.sign(&digest);
-    let envelope = json!({
-        "key_id": "http-test-publisher",
-        "signature": base64::engine::general_purpose::STANDARD.encode(signature.to_bytes()),
-    })
-    .to_string();
-
-    let file = std::fs::File::create(path).expect("create signed zip");
-    let mut zip = zip::ZipWriter::new(file);
-    let opts = zip::write::SimpleFileOptions::default()
-        .compression_method(zip::CompressionMethod::Deflated);
-    for (name, body) in entries {
-        zip.start_file(name, opts).expect("start signed entry");
-        std::io::Write::write_all(&mut zip, body.as_bytes()).expect("write signed entry");
-    }
-    zip.start_file("tandempack.sig", opts)
-        .expect("start pack signature");
-    std::io::Write::write_all(&mut zip, envelope.as_bytes()).expect("write pack signature");
-    zip.finish().expect("finish signed zip");
-
-    let previous = std::env::var("TANDEM_PACK_TRUSTED_PUBLIC_KEYS").ok();
-    std::env::set_var(
-        "TANDEM_PACK_TRUSTED_PUBLIC_KEYS",
-        format!(
-            "http-test-publisher={}",
-            base64::engine::general_purpose::STANDARD
-                .encode(signing_key.verifying_key().to_bytes())
-        ),
-    );
-    TrustedPackKeyGuard { previous }
-}
 
 fn direct_loopback() -> axum::extract::ConnectInfo<std::net::SocketAddr> {
     axum::extract::ConnectInfo("127.0.0.1:39731".parse().expect("loopback socket address"))
